#72 #73 complete account register api development.

+++ backend/src/main/java/com/amazon/entity/CurrentUser.java
@@ -8,14 +8,14 @@ import java.util.Collection;
 
 public class CurrentUser extends User implements Serializable {
 
-    private final int accountId;
+    private final long accountId;
 
     public CurrentUser(Account account, Collection<? extends GrantedAuthority> authorities) {
         super(account.getEmail(), account.getPassword(), authorities);
         this.accountId = account.getId();
     }
 
-    public int getAccountId(){
+    public long getAccountId(){
         return accountId;
     }
